chore: integrate all modules, fix compilation, 91 tests pass
Pull fully integrated block_device.rs, cache.rs, fuse_fs.rs from
integration worktree. All 91 tests pass. Release build succeeds.
Binary creates valid TOML template on first run.

+++ src/fuse_fs.rs
@@ -1,7 +1,6 @@
 use std::ffi::OsStr;
 use std::time::{Duration, SystemTime};
 
-use bytes::Bytes;
 use fuser::{
     BsdFileFlags, Config, Errno, FileAttr, FileHandle, FileType, Filesystem, FopenFlags,
     Generation, INodeNo, KernelConfig, LockOwner, MountOption, OpenFlags, ReplyAttr, ReplyData,
@@ -61,7 +60,7 @@ impl IscsiFuseFs {
 
         let mut config = Config::default();
         config.mount_options = options;
-        config.n_threads = Some(num_cpus::get() as u32);
+        config.n_threads = Some(num_cpus::get());
         config
     }
